requirements.txt
faiss-cpu
fastapi
google-generativeai
markdownify
numpy
pillow
playwright
python-dotenv
requests
sentence-transformers
torch
transformers
uvicorn
uvloop
wheel
